+++ README.md
@@ -32,50 +32,63 @@ Install-Package CohereDotnet
 
 3. **Example Usage**:
 
-- **Generate Text**:
+- **Chat Endpoint**:
+
+    Use the ```Chat``` endpoint to generate conversational text by sending a series of messages. The [ChatRequest](./Cohere/Types/ChatRequest.cs) object allows you to customize settings for tailored responses.
 
     ```csharp
-    var response = await cohereClient.ChatAsync(new GenerateRequest
+    using Cohere;
+    using Cohere.Types;
+
+    var response = await cohereClient.ChatAsync(new ChatRequest
     {
-        Messages = [
-            {
-                "role": "user",
-                "content": "hello world!"
-            }
-        ]
+        Messages = new List<ChatMessage>
+        {
+            new ChatMessage { Role = "user", Content = "Hello, how can I get started with Cohere?" }
+        },
         MaxTokens = 50
     });
-    Console.WriteLine(response.FinishReason);
+
+    var responseContent = response.Message?.Content as List<ChatResponseMessageText>;
+
+    // Print model's response text
+    Console.WriteLine(responseContent?[0].Text);
     ```
 
-- **Classify Text**:
+- **Classify Endpoint**:
+
+    Classify text inputs into specified labels using the ```Classify``` endpoint. The [ClassifyRequest](./Cohere/Types/ClassifyRequest.cs) object allows you to customize settings for tailored responses.
 
     ```csharp
     var response = await cohereClient.ClassifyAsync(new ClassifyRequest
     {
-        Examples = new List<(string text, string label)>
-        {
-            (text="Dermatologists don't like her!", label="Spam"),
-            (text="'Hello, open to this?'", label="Spam"),
-            (text="Your parcel will be delivered today", label="Not spam"),
-            (text="Review changes to our Terms and Conditions", label="Not spam"),
-        };
-        Inputs = new List<string>
-        {
+        Examples =
+        [
+            new ClassifyExample { Text = "Dermatologists don't like her!", Label = "Spam" },
+            new ClassifyExample { Text = "'Hello, open to this?'", Label = "Spam" },
+            new ClassifyExample { Text = "Your parcel will be delivered today", Label = "Not spam" },
+            new ClassifyExample { Text = "Review changes to our Terms and Conditions", Label = "Not spam" },
+        ],
+        Inputs =
+        [
             "Confirm your email address",
             "hey i need u to send some $"
-        };
+        ]
     });
-    Console.WriteLine(response.Classifications.First().Prediction);
+
+    // Print label prediction of first input
+    Console.WriteLine(response.Classifications[0].Predictions[0]);
     ```
 
-- **Rerank Text**:
+- **Rerank Endpoint**:
+
+    Use the ```Rerank``` endpoint to arrange a list of documents based on relevance to a given query. The [RerankRequest](./Cohere/Types/RerankRequest.cs) object allows you to customize settings for tailored responses.
 
     ```csharp
     var response = await cohereClient.RerankAsync(new RerankRequest
     {
         Query = "What is the capital of the United States?"
-        Documents = new List<string>
+        Documents = new List<object>
         {
             "Carson City is the capital city of the American state of Nevada.",
             "The Commonwealth of the Northern Mariana Islands is a group of islands in the Pacific Ocean. Its capital is Saipan.",
@@ -84,9 +97,26 @@ Install-Package CohereDotnet
             "Capital punishment (the death penalty) has existed in the United States since beforethe United States was a country. As of 2017, capital punishment is legal in 30 of the 50 states.",
         }
     });
-    foreach (var result in Results)
+
+    // Print index and relevance score of each document
+    foreach (var result in response.Results)
     {
-        Console.WriteLine(result.ToString());
+        Console.WriteLine($"Index: {results.Index}, Relevance Score: {result.RelevanceScore}");
+    }
+    ```
+
+4. **Error Handling**:
+
+    The CohereClient class includes error handling to ensure clear feedback when requests fail:
+
+    ```csharp
+    try
+    {
+        var response = await cohereClient.ChatAsync(new ChatRequest { Messages = null });
+    }
+    catch (CohereApiException ex)
+    {
+        Console.WriteLine(ex.ToString());
     }
     ```
